refactor: Typescript example
